src/analyze-component/requirements.txt
awsiotsdk
opencv-python
boto3